+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- Added RotatingFileHandler and StreamHandler to log both to a file and stdout
+- 
 
 ## [0.0.2-alpha] - 2022-03-19